+++ bot/src/services/API/subU.service.ts
@@ -1,10 +1,14 @@
 import axios from "axios";
 import { config } from "../config/config.service.js";
 
-export const deposit10 = async () => {
-  return await axios.get(`${config.get("SUBU_URL")}/api/deposit/10`, {
-    headers: { Authorization: `user ${config.get("JWT_TOKEN")}` },
-  });
+export const deposit10 = async (toUsername: string) => {
+  return await axios.post(
+    `${config.get("SUBU_URL")}/api/deposit/10`,
+    { toUser: toUsername },
+    {
+      headers: { Authorization: `user ${config.get("JWT_TOKEN")}` },
+    }
+  );
 };
 
 export const checkJWT = async () => {