+++ server/src/services/config/config.interface.ts
@@ -1,3 +1,4 @@
 export interface ConfigServiceI {
-  get(key: string): string;
+  get(key: string): string | undefined;
+  set(key: string, value: string): void;
 }