+++ server/src/services/config/config.service.ts
@@ -1,25 +1,27 @@
-import { config, DotenvParseOutput } from "dotenv";
-import { ConfigServiceI } from "./config.interface";
+import { ConfigServiceI } from "./config.interface.js";
+import { config as dotenvConfig, DotenvParseOutput } from "dotenv";
 
-export class ConfigService implements ConfigServiceI {
+class ConfigService implements ConfigServiceI {
   private config: DotenvParseOutput;
 
   constructor() {
-    const { error, parsed } = config();
+    const { error, parsed: parsedConfig } = dotenvConfig();
 
     if (error) throw error;
-    if (!parsed) throw new Error(".env is empty");
+    if (!parsedConfig) throw new Error(".env is empty");
 
-    this.config = parsed;
+    this.config = parsedConfig;
   }
 
-  get(key: string): string {
+  get(key: string) {
     const res = this.config[key];
 
-    if (!res) {
-      throw new Error(`Key "${key}" not found in.env`);
-    }
+    if (res) return res;
+  }
 
-    return res;
+  set(key: string, value: string) {
+    this.config[key] = value;
   }
 }
+
+export const config = new ConfigService();